+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Au.Consumer/Models/StudentPersonal.cs
@@ -21,8 +21,8 @@ using System.Xml.Serialization;
 namespace Sif.Framework.Demo.Au.Consumer.Models
 {
 
-    [XmlRoot("StudentPersonal", Namespace = "http://www.sifassociation.org/au/datamodel/1.4", IsNullable = false)]
-    [XmlType(Namespace = "http://www.sifassociation.org/au/datamodel/1.4")]
+    [XmlRoot("StudentPersonal", Namespace = "http://www.sifassociation.org/au/datamodel/3.4", IsNullable = false)]
+    [XmlType(Namespace = "http://www.sifassociation.org/au/datamodel/3.4")]
     public class StudentPersonal : StudentPersonalType, IDataModel
     {
     }
