+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Au.Consumer/SifFramework.config
@@ -6,7 +6,6 @@
     </sectionGroup>
   </configSections>
   <appSettings>
-
     <!-- Settings used to register this Consumer with the Sif3Framework EnvironmentProvider. -->
     <add key="consumer.environment.deleteOnUnregister" value="true"/>
     <add key="consumer.environment.sharedSecret" value="SecretDem0"/>
@@ -14,31 +13,14 @@
     <add key="consumer.environment.template.applicationKey" value="Sif3DemoApp"/>
     <add key="consumer.environment.template.authenticationMethod" value="Basic"/>
     <add key="consumer.environment.template.consumerName" value="DemoConsumer"/>
-    <add key="consumer.environment.template.dataModelNamespace" value="http://www.sifassociation.org/au/datamodel/1.4"/>
-    <add key="consumer.environment.template.supportedInfrastructureVersion" value="3.0.1"/>
+    <add key="consumer.environment.template.dataModelNamespace" value="http://www.sifassociation.org/au/datamodel/3.4"/>
+    <add key="consumer.environment.template.solutionId" value="Sif3Framework"/>
+    <add key="consumer.environment.template.supportedInfrastructureVersion" value="3.2"/>
     <add key="consumer.paging.navigationPageSize" value="5"/>
-
-    <!-- Settings used to register this Consumer with the HITS EnvironmentProvider. -->
-    <!--
-    <add key="consumer.environment.deleteOnUnregister" value="false"/>
-    <add key="consumer.environment.sharedSecret" value="hits@nsip"/>
-    <add key="consumer.environment.url" value="http://hits.dev.nsip.edu.au/SIF3InfraREST/hits/environments/environment"/>
-    <add key="consumer.environment.template.applicationKey" value="HITS"/>
-    <add key="consumer.environment.template.authenticationMethod" value="Basic"/>
-    <add key="consumer.environment.template.consumerName" value=""/>
-    <add key="consumer.environment.template.dataModelNamespace" value="http://www.sifassociation.org/au/datamodel/1.4"/>
-    <add key="consumer.environment.template.solutionId" value="HITS"/>
-    <add key="consumer.environment.template.supportedInfrastructureVersion" value="3.0.1"/>
-    <add key="consumer.environment.template.userToken" value="0EE41AE6-C43F-11E3-9050-E0F4DBD909AB"/>
-    <add key="consumer.paging.navigationPageSize" value="5"/>
-    -->
   </appSettings>
   <sifFramework>
     <consumer>
       <sessions>
-        <!-- HITS pre-defined authorisation information required for Consumer requests. -->
-        <session applicationKey="HITS" environmentUrl="http://hits.dev.nsip.edu.au/SIF3InfraREST/hits/environments/4482c242-7945-40e6-a436-7290b9830ad1"
-          sessionToken="79623ef1-98c5-4dd0-802e-822589b78b81" solutionId="HITS" userToken="0EE41AE6-C43F-11E3-9050-E0F4DBD909AB" />
       </sessions>
     </consumer>
   </sifFramework>
