+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Au.Provider/SifFramework.config
@@ -13,7 +13,7 @@
     <add key="provider.environment.template.applicationKey" value="Sif3DemoApp"/>
     <add key="provider.environment.template.authenticationMethod" value="Basic"/>
     <add key="provider.environment.template.consumerName" value="DemoProvider"/>
-    <add key="provider.environment.template.dataModelNamespace" value="http://www.sifassociation.org/au/datamodel/1.4"/>
+    <add key="provider.environment.template.dataModelNamespace" value="http://www.sifassociation.org/au/datamodel/3.4"/>
     <add key="provider.environment.template.supportedInfrastructureVersion" value="3.2"/>
   </appSettings>
   <sifFramework>
