+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Hits.Consumer/App.config
@@ -1,0 +1,70 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+
+  <!-- Register a section handler for the log4net section -->
+  <configSections>
+    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
+  </configSections>
+
+  <!-- This section contains the log4net configuration settings. -->
+  <log4net>
+    <appender name="ManagedColoredConsoleAppender" type="log4net.Appender.ManagedColoredConsoleAppender">
+      <layout type="log4net.Layout.PatternLayout">
+        <conversionPattern value="%date{yyyyMMdd;HH:mm:ss} %level %logger{1} - %message%newline"/>
+      </layout>
+      <mapping>
+        <level value="ERROR" />
+        <foreColor value="DarkRed" />
+      </mapping>
+      <mapping>
+        <level value="WARN" />
+        <foreColor value="DarkYellow" />
+      </mapping>
+      <mapping>
+        <level value="DEBUG" />
+        <foreColor value="DarkGreen" />
+      </mapping>
+      <mapping>
+        <level value="INFO" />
+        <foreColor value="DarkBlue" />
+      </mapping>
+      <layout type="log4net.Layout.PatternLayout">
+        <conversionPattern value="%date [%thread] %-5level %logger [%property{NDC}] - %message%newline" />
+      </layout>
+    </appender>
+    <appender name="RollingLogFileAppender" type="log4net.Appender.RollingFileAppender">
+      <file value="App.log"/>
+      <appendToFile value="true"/>
+      <lockingModel type="log4net.Appender.FileAppender+MinimalLock"/>
+      <maxSizeRollBackups value="10"/>
+      <maximumFileSize value="1MB"/>
+      <rollingStyle value="Size"/>
+      <staticLogFileName value="true"/>
+      <layout type="log4net.Layout.PatternLayout">
+        <header value="[START %utcdate]&#xD;&#xA;" />
+        <footer value="[END   %utcdate]&#xD;&#xA;" />
+        <conversionPattern value="%utcdate %-5level %logger{3} - %property{instance} - %message%newline"/>
+      </layout>
+    </appender>
+    <root>
+      <level value="WARN"/>
+      <appender-ref ref="ManagedColoredConsoleAppender"/>
+      <appender-ref ref="RollingLogFileAppender"/>
+    </root>
+    <!-- Specify the level for some specific categories -->
+    <!--logger name="NHibernate">
+      <level value="WARN"/>
+    </logger-->
+    <logger name="Sif.Framework">
+      <level value="DEBUG"/>
+    </logger>
+    <logger name="Sif.Framework.Demo.Au.Consumer">
+      <level value="DEBUG"/>
+    </logger>
+  </log4net>
+
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+  </startup>
+
+</configuration>
