+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Hits.Consumer/SifFramework.config
@@ -1,0 +1,31 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <configSections>
+    <sectionGroup name="sifFramework" type="Sif.Framework.Model.Settings.SifFrameworkSectionGroup, Sif.Framework">
+      <section name="consumer" type="Sif.Framework.Model.Settings.ConsumerSection, Sif.Framework" />
+    </sectionGroup>
+  </configSections>
+  <appSettings>
+    <!-- Settings used to register this Consumer with the HITS EnvironmentProvider. -->
+    <add key="consumer.environment.deleteOnUnregister" value="false"/>
+    <add key="consumer.environment.sharedSecret" value="bcbb4c5d-3338-4785-a63c-7add2194894c"/>
+    <add key="consumer.environment.url" value="http://hits.dev.nsip.edu.au/SIF3InfraREST/hits/environments/environment"/>
+    <add key="consumer.environment.template.applicationKey" value="614bebe2-c529-4ef8-8a10-d04e18facc40"/>
+    <add key="consumer.environment.template.authenticationMethod" value="Basic"/>
+    <add key="consumer.environment.template.consumerName" value=""/>
+    <add key="consumer.environment.template.dataModelNamespace" value="http://www.sifassociation.org/au/datamodel/3.4"/>
+    <add key="consumer.environment.template.solutionId" value="HITS"/>
+    <add key="consumer.environment.template.supportedInfrastructureVersion" value="3.1"/>
+    <add key="consumer.environment.template.userToken" value="6dd11373-95fd-44d0-84b5-f6a238d78a6a"/>
+    <add key="consumer.paging.navigationPageSize" value="5"/>
+  </appSettings>
+  <sifFramework>
+    <consumer>
+      <sessions>
+        <!-- Pre-defined authorisation values required for Consumer requests where an Environment in HITS already exists. -->
+        <session applicationKey="614bebe2-c529-4ef8-8a10-d04e18facc40" environmentUrl="http://hits.dev.nsip.edu.au/SIF3InfraREST/hits/environments/cbde5abb-bef7-4132-bc3a-58db4d023178"
+          sessionToken="04fea921-a651-4fa6-98ba-8d7684835367" solutionId="HITS" userToken="6dd11373-95fd-44d0-84b5-f6a238d78a6a" />
+      </sessions>
+    </consumer>
+  </sifFramework>
+</configuration>
