+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Hits.Consumer/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="AutoMapper" version="3.1.1" targetFramework="net45" />
+  <package id="log4net" version="2.0.3" targetFramework="net45" />
+</packages>
