+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Us.Consumer/Consumers/XStudentConsumer.cs
@@ -14,17 +14,26 @@
  * limitations under the License.
  */
 
-using Sif.Framework.Model.DataModels;
-using Sif.Specification.DataModel.Au;
-using System.Xml.Serialization;
+using Sif.Framework.Consumers;
+using Sif.Framework.Demo.Us.Consumer.Models;
+using Sif.Framework.Model.Infrastructure;
 
-namespace Sif.Framework.Demo.Au.Consumer.Models
+namespace Sif.Framework.Demo.Us.Consumer.Consumers
 {
 
-    [XmlRoot("StaffPersonal", Namespace = "http://www.sifassociation.org/au/datamodel/1.4", IsNullable = false)]
-    [XmlType(Namespace = "http://www.sifassociation.org/au/datamodel/1.4")]
-    public class StaffPersonal : StaffPersonalType, IDataModel
+    class XStudentConsumer : BasicConsumer<XStudent>
     {
+
+        public XStudentConsumer(string applicationKey, string instanceId = null, string userToken = null, string solutionId = null)
+            : base(applicationKey, instanceId, userToken, solutionId)
+        {
+        }
+
+        public XStudentConsumer(Environment environment)
+            : base(environment)
+        {
+        }
+
     }
 
 }
