+++ Code/Sif3FrameworkDemo/Sif.Framework.Demo.Us.Consumer/Sif.Framework.Demo.Us.Consumer.csproj
@@ -65,7 +65,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="XStudentConsumer.cs" />
+    <Compile Include="Consumers\XStudentConsumer.cs" />
     <Compile Include="ConsumerApp.cs" />
     <Compile Include="Models\XStudent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
