+++ SharedLibs/Sif.Framework 1.0.0/README.txt
@@ -1,0 +1,1 @@
+Core library of the SIF3 Framework based on SIF Infrastructure 3.0.1.
